python_backend/requirements.txt
fastapi
pydantic
uvicorn
requests
python-dotenv>=1.0.0